Menus and Shortcuts: split out the Win32 menu support
This extracts the menu logic out of the various files into a standalone
platform support structure to isolate the logic.
